ImportStatement for automatic import insertion is now an interface.
This way, it can be implemented by other kinds of completions,
not just a BasicCompletion.
