+++ src/main/java/org/scijava/ui/swing/script/autocompletion/ImportCompletion.java
@@ -1,49 +1,16 @@
-/*-
- * #%L
- * Script Editor and Interpreter for SciJava script languages.
- * %%
- * Copyright (C) 2009 - 2020 SciJava developers.
- * %%
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- * 
- * 1. Redistributions of source code must retain the above copyright notice,
- *    this list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- * 
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDERS OR CONTRIBUTORS BE
- * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
- * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
- * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
- * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- * #L%
- */
 package org.scijava.ui.swing.script.autocompletion;
 
-import org.fife.ui.autocomplete.BasicCompletion;
-import org.fife.ui.autocomplete.CompletionProvider;
-
-public class ImportCompletion extends BasicCompletion
+public interface ImportCompletion
 {
-	protected final String importStatement,
-						   className;
+	/**
+	 * 
+	 * @return The formatted import statement, e.g. "from this import that".
+	 */
+	public String getImportStatement();
 	
-	public ImportCompletion(final CompletionProvider provider, final String replacementText, final String className, final String importStatement) {
-		super(provider, replacementText);
-		this.className = className;
-		this.importStatement = importStatement;
-	}
-
-	@Override
-	public String getSummary() {
-		return importStatement;
-	}
+	/**
+	 * 
+	 * @return The fully qualified class name.
+	 */
+	public String getClassName();
 }
